Add useFirebaseAuthListener and refactor auth usage
Introduce a centralized useFirebaseAuthListener hook (app/NX/lib/hooks/useFirebaseAuthListener.ts) that sets global paywall state and notifies callers. Replace scattered onAuthStateChanged logic in Settings, RequireAuth, useAuthed, and NXAdminAuthWrapper with the new hook to eliminate duplication and simplify auth handling. Add a client-side lib index (app/NX/lib/index.ts) exporting utilities including useFirebaseAuthListener, and move server-only exports to app/NX/lib/index.server.ts, updating imports across pages to reference index.server where appropriate. Also fix a getMeta import path and perform minor cleanups (remove unused imports, formatting).
