+++ app/NX/Paywall/components/RequireAuth.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import type { T_Config } from "../../types";
 import SignIn from './SignIn';
 import { firebaseLogin } from '../../Paywall';
-import { getFirebaseAuth } from '../../lib/firebase';
-import { onAuthStateChanged, User } from 'firebase/auth';
+import { User } from 'firebase/auth';
+import { useFirebaseAuthListener } from '../../lib';
 import { Typography, Backdrop, CircularProgress, Box } from "@mui/material";
 import { useDispatch } from '../../Uberedux';
 import { setPaywall } from '../../Paywall';
@@ -25,35 +25,10 @@ export default function RequireAuth({ children, config }: { children: React.Reac
         }
     };
 
-    useEffect(() => {
-        const auth = getFirebaseAuth();
-        const unsubscribe = onAuthStateChanged(auth, (firebaseUser) => {
-            setUser(firebaseUser);
-            let safeUser = null;
-            if (firebaseUser) {
-                const { uid, email, emailVerified, isAnonymous, providerData, displayName, photoURL } = firebaseUser;
-                safeUser = {
-                    uid,
-                    email,
-                    emailVerified,
-                    isAnonymous,
-                    providerData: providerData?.map(p => ({
-                        providerId: p.providerId,
-                        uid: p.uid,
-                        displayName: p.displayName,
-                        email: p.email,
-                        phoneNumber: p.phoneNumber,
-                        photoURL: p.photoURL
-                    })),
-                    displayName: displayName ?? null,
-                    photoURL: photoURL ?? null
-                };
-            }
-            dispatch(setPaywall("user", safeUser));
-            setLoading(false);
-        });
-        return () => unsubscribe();
-    }, []);
+    useFirebaseAuthListener((firebaseUser) => {
+        setUser(firebaseUser);
+        setLoading(false);
+    });
 
     if (loading) return (
         <Backdrop open sx={{ zIndex: (theme) => theme.zIndex.drawer + 1, color: '#fff' }}>
