+++ app/NX/lib/index.server.ts
@@ -1,20 +1,15 @@
-import { getTenant } from './getTenant';
 import { serverUseNav } from './serverHooks/serverUseNav';
 import { serverUseConfig } from './serverHooks/serverUseConfig';
 import { serverUseMDBySlug } from './serverHooks/serverUseMDBySlug';
 import { serverUseAllMd } from './serverHooks/serverUseAllMd';
-import { createSlug } from './vanilla-js/createSlug';
-import { makeIdentity } from './vanilla-js/makeIdentity';
-import { militaryTime } from './vanilla-js/militaryTime';
+import { getTenant } from './getTenant';
 import { getMeta } from './getMeta';
+
 export {
     serverUseNav,
     serverUseConfig,
     serverUseMDBySlug,
     serverUseAllMd,
-    createSlug,
     getTenant,
     getMeta,
-    militaryTime,
-    makeIdentity,
 };
