+++ app/NX/lib/index.ts
@@ -1,0 +1,11 @@
+import { createSlug } from './vanilla-js/createSlug';
+import { makeIdentity } from './vanilla-js/makeIdentity';
+import { militaryTime } from './vanilla-js/militaryTime';
+import { useFirebaseAuthListener } from './hooks/useFirebaseAuthListener';
+
+export {
+    createSlug,
+    militaryTime,
+    useFirebaseAuthListener,
+    makeIdentity,
+};
